+++ src/users/JuniorProducerClass.java
@@ -13,7 +13,7 @@ public class JuniorProducerClass extends ProducerClass implements Producer {
 	 * @param hourlyCost junior producer's cost charged per hour.
 	 * @param username name of the junior producer.
 	 */
-	public JuniorProducerClass(int hourlyCost, String username) {
+	public JuniorProducerClass(float hourlyCost, String username) {
 		super(hourlyCost, username);
 	}
