+++ src/users/UserClass.java
@@ -17,7 +17,7 @@ public abstract class UserClass implements User {
 	/**
 	 * Collaborator's cost charged per hour.
 	 */
-	private int hourlyCost;
+	private float hourlyCost;
 	
 	/**
 	 * Collaborator's calendar with scheduled recordings dates.
@@ -29,7 +29,7 @@ public abstract class UserClass implements User {
 	 * @param hourlyCost cost per hour charged by the collaborator.
 	 * @param username collaborator's name.
 	 */
-	protected UserClass(int hourlyCost, String username) {
+	protected UserClass(float hourlyCost, String username) {
 		this.username = username;
 		this.hourlyCost = hourlyCost;
 		calendar = new ArrayClass<LocalDateTime>();	//with default size of 50
@@ -40,7 +40,7 @@ public abstract class UserClass implements User {
 		return username;
 	}
 	
-	public int getHourlyCost() {
+	public float getHourlyCost() {
 		return hourlyCost;
 	}
